+++ edison-mongo/src/main/java/de/otto/edison/mongo/configuration/MongoJobsConfiguration.java
@@ -8,6 +8,7 @@ import de.otto.edison.mongo.jobs.MongoJobRepository;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -15,24 +16,29 @@ import static org.slf4j.LoggerFactory.getLogger;
 
 @Configuration
 @ConditionalOnClass(name = "de.otto.edison.jobs.configuration.JobsConfiguration")
+@EnableConfigurationProperties(MongoProperties.class)
 public class MongoJobsConfiguration {
 
     private static final Logger LOG = getLogger(MongoJobsConfiguration.class);
 
     @Bean
-    public JobRepository jobRepository(final MongoDatabase mongoDatabase, final @Value("${edison.jobs.collection.jobinfo:jobinfo}") String collectionName) {
+    public JobRepository jobRepository(final MongoDatabase mongoDatabase,
+                                       final @Value("${edison.jobs.collection.jobinfo:jobinfo}") String collectionName,
+                                       final MongoProperties mongoProperties) {
         LOG.info("===============================");
         LOG.info("Using MongoJobRepository with {} MongoDatabase impl.", mongoDatabase.getClass().getSimpleName());
         LOG.info("===============================");
-        return new MongoJobRepository(mongoDatabase, collectionName);
+        return new MongoJobRepository(mongoDatabase, collectionName, mongoProperties);
     }
 
     @Bean
-    public JobMetaRepository jobMetaRepository(final MongoDatabase mongoDatabase, final @Value("${edison.jobs.collection.jobmeta:jobmeta}") String collectionName) {
+    public JobMetaRepository jobMetaRepository(final MongoDatabase mongoDatabase,
+                                               final @Value("${edison.jobs.collection.jobmeta:jobmeta}") String collectionName,
+                                               final MongoProperties mongoProperties) {
         LOG.info("===============================");
         LOG.info("Using MongoJobMetaRepository with {} MongoDatabase impl.", mongoDatabase.getClass().getSimpleName());
         LOG.info("===============================");
-        return new MongoJobMetaRepository(mongoDatabase, collectionName);
+        return new MongoJobMetaRepository(mongoDatabase, collectionName, mongoProperties);
     }
 
 }